requirements.txt
langchain
langchain-groq
streamlit
python-dotenv
youtube-transcript-api
